Added a linter and linted files

+++ src/commands/moderation/purge.js
@@ -1,36 +1,34 @@
 module.exports = {
 	async run(message, args, colors) {
-		if(args.length == 1)
-			message.channel.bulkDelete(2);
+		if (args.length == 1) message.channel.bulkDelete(2);
 		else {
 			const num = parseInt(args[1]);
-			if(isNaN(num))
-			{
+			if (isNaN(num)) {
 				return message.channel.send('', {
 					embed: {
 						title: 'Please enter a valid number.',
-						color: colors.error
-					}
+						color: colors.error,
+					},
 				});
 			}
-			if(num > 99) {
+			if (num > 99) {
 				message.channel.send('', {
 					embed: {
 						title: 'Please enter a smaller number.',
-						color: colors.error
-					}
+						color: colors.error,
+					},
 				});
 			} else {
-				return message.channel.bulkDelete(num+1, true)
-				.then(messages => 
-					message.channel.send({
-						embed: {
-							description: `Bulk deleted ${messages.size} messages`,
-							color: colors.success
-						}
-					})
-					.then(msg => msg.delete({timeout: 3500}))
-					.catch()
+				return message.channel.bulkDelete(num + 1, true).then((messages) =>
+					message.channel
+						.send({
+							embed: {
+								description: `Bulk deleted ${messages.size} messages`,
+								color: colors.success,
+							},
+						})
+						.then((msg) => msg.delete({ timeout: 3500 }))
+						.catch()
 				);
 			}
 		}
@@ -38,9 +36,9 @@ module.exports = {
 	aliases: ['p'],
 	description: 'Purges Messages',
 	detailed: 'Purges Messages',
-	examples: prefix => `${prefix}purge <number of message to delete>`,
+	examples: (prefix) => `${prefix}purge <number of message to delete>`,
 	name: 'purge',
 	permissions: ['MANAGE_MESSAGES'],
 	botPermissions: ['MANAGE_MESSAGES'],
-	guildOnly: true
-}
+	guildOnly: true,
+};
